\name{print.bayes.factor}
\alias{print.bayes.factor}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ Printing for Bayes Factor Objects }
\description{
  Prints a quick summary of a Bayes Factor object.
}
\usage{
print.bayes.factor(bf)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{bf}{ An object of class \code{bayes.factor} }
}
%\details{
%}
\value{
None
}
%\references{  }
\author{ Carter T. Butts \email{ctb@andrew.cmu.edu} }
%\note{ }

% ~Make other sections like WARNING with \section{WARNING }{....} ~

\seealso{ \code{\link{bbnam.bf}}}

\examples{
}
\keyword{ print }%-- one or more ...
\keyword{ network }