\name{print.bbnam}
\alias{print.bbnam}
\alias{print.bbnam.fixed}
\alias{print.bbnam.pooled}
\alias{print.bbnam.actor}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ Printing for bbnam Objects }
\description{
Prints a quick summary of posterior draws from \code{\link{bbnam}}.
}
\usage{
print.bbnam(b)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{b}{ A \code{\link{bbnam}} object }
}
%\details{
%}
\value{
None
}
%\references{ }
\author{ Carter T. Butts \email{ctb@andrew.cmu.edu} }
%\note{ }

% ~Make other sections like WARNING with \section{WARNING }{....} ~

\seealso{ \code{\link{bbnam}} }

\examples{
}
\keyword{ print }%-- one or more ...
\keyword{ network }