\name{print.blockmodel}
\alias{print.blockmodel}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ Printing for blockmodel Objects }
\description{
Prints a quick summary of a \code{\link{blockmodel}} object.
}
\usage{
\method{print}{blockmodel}(x, ...)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{x}{ An object of class \code{blockmodel} }
  \item{\dots}{ Further arguments passed to or from other methods }
}
%\details{
%}
\value{
None
}
%\references{ }
\author{ Carter T. Butts \email{buttsc@uci.edu} }
%\note{  }

% ~Make other sections like WARNING with \section{WARNING }{....} ~

\seealso{ \code{\link{blockmodel}} }

%\examples{
%}
\keyword{ print }%-- one or more ...
