\name{print.cugtest}
\alias{print.cugtest}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ Printing for cugtest Objects}
\description{
  Prints a quick summary of objects produced by \code{\link{cugtest}}.
}
\usage{
print.cugtest(c)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{c}{ An object of class \code{cugtest} }
}
%\details{
%}
\value{
None.
}
%\references{}
\author{ Carter T. Butts \email{ctb@andrew.cmu.edu} }
%\note{  }

% ~Make other sections like WARNING with \section{WARNING }{....} ~

\seealso{ \code{\link{cugtest}} }

\examples{
}
\keyword{ print }%-- one or more ...
\keyword{ network }