\name{print.netcancor}
\alias{print.netcancor}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ Printing for netcancor Objects }
\description{
  Prints a quick summary of objects produced by \code{\link{netcancor}}.
}
\usage{
print.netcancor(n)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{n}{ An object of class \code{netcancor}}
}
%\details{
%}
\value{
}
%\references{}
\author{ Carter T. Butts \email{ctb@andrew.cmu.edu} }
%\note{  }

% ~Make other sections like WARNING with \section{WARNING }{....} ~

\seealso{ \code{\link{netcancor}} }

\examples{
}
\keyword{ print }%-- one or more ...
\keyword{ network }