\name{print.netlm}
\alias{print.netlm}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ Printing for netlm Objects }
\description{
  Prints a quick summary of objects produced by \code{\link{netlm}}.
}
\usage{
print.netlm(n)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{n}{ An object of class \code{netlm}}
}
%\details{
%}
\value{
}
%\references{  }
\author{ Carter T. Butts \email{ctb@andrew.cmu.edu} }
%\note{  }

 %~Make other sections like WARNING with \section{WARNING }{....} ~

\seealso{ \code{\link{netlm}}}

\examples{
}
\keyword{ print }%-- one or more ...
\keyword{ network }