\name{print.netlogit}
\alias{print.netlogit}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ Printing for netlogit Objects }
\description{
  Prints a quick summary of objects produced by \code{\link{netlogit}}.
}
\usage{
print.netlogit(n)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{n}{ An object of class \code{netlogit} }
}
%\details{
%}
\value{
}
%\references{  }
\author{ Carter T. Butts \email{ctb@andrew.cmu.edu} }
%\note{  }

% ~Make other sections like WARNING with \section{WARNING }{....} ~

\seealso{ \code{\link{netlogit}} }

\examples{
}
\keyword{ print }%-- one or more ...
\keyword{ network }