\name{print.qaptest}
\alias{print.qaptest}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ Printing for qaptest Objects }
\description{
Prints a quick summary of objects produced by \code{\link{qaptest}}.
}
\usage{
\method{print}{qaptest}(x, ...)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{x}{ An object of class \code{qaptest} }
  \item{\dots}{ Further arguments passed to or from other methods }
}
%\details{
%}
%\value{
%}
%\references{  }
\author{ Carter T. Butts \email{buttsc@uci.edu} }
%\note{  }

% ~Make other sections like WARNING with \section{WARNING }{....} ~

\seealso{ \code{\link{qaptest}}}

%\examples{
%}
\keyword{ print }%-- one or more ...
