\name{print.summary.bbnam}
\alias{print.summary.bbnam}
\alias{print.summary.bbnam.fixed}
\alias{print.summary.bbnam.pooled}
\alias{print.summary.bbnam.actor}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ Printing for summary.bbnam Objects }
\description{
  Prints an object of class \code{summary.bbnam}.
}
\usage{
print.summary.bbnam(b)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{b}{ An object of class \code{summary.bbnam} }
}
%\details{
%}
\value{
}
%\references{ }
\author{ Carter T. Butts \email{ctb@andrew.cmu.edu} }
%\note{  }

% ~Make other sections like WARNING with \section{WARNING }{....} ~

\seealso{ \code{\link{bbnam}} }

\examples{

}
\keyword{ print }%-- one or more ...
\keyword{ network }