\name{print.summary.blockmodel}
\alias{print.summary.blockmodel}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ Printing for summary.blockmodel Objects }
\description{
  Prints an object of class \code{summary.blockmodel}.
}
\usage{
print.summary.blockmodel(b)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{b}{ An object of class \code{summary.blockmodel} }
}
%\details{
%}
\value{
}
%\references{  }
\author{ Carter T. Butts \email{ctb@andrew.cmu.edu} }
%\note{ }

% ~Make other sections like WARNING with \section{WARNING }{....} ~

\seealso{ \code{\link{summary.blockmodel}}}

\examples{

}
\keyword{ print }%-- one or more ...
\keyword{ network }