\name{print.summary.cugtest}
\alias{print.summary.cugtest}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ Printing for summary.cugtest Objects }
\description{
  Prints an object of class \code{summary.cugtest}.
}
\usage{
print.summary.cugtest(c)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{c}{ An object of class \code{summary.cugtest} }
}
%\details{
%}
\value{
}
%\references{  }
\author{ Carter T. Butts \email{ctb@andrew.cmu.edu} }
%\note{  }

% ~Make other sections like WARNING with \section{WARNING }{....} ~

\seealso{ \code{\link{summary.cugtest}} }

\examples{

}
\keyword{ print }%-- one or more ...
\keyword{ network }