\name{print.summary.netcancor}
\alias{print.summary.netcancor}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ Printing for summary.netcancor Objects }
\description{
  Prints an object of class \code{summary.netcancor}.
}
\usage{
print.summary.netcancor(n)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{n}{ An object of class \code{summary.netcancor}}
}
%\details{
%}
\value{
}
%\references{ }
\author{ Carter T. Butts \email{ctb@andrew.cmu.edu} }
%\note{  }

 %~Make other sections like WARNING with \section{WARNING }{....} ~

\seealso{ \code{\link{summary.netcancor}} }

\examples{

}
\keyword{ print }%-- one or more ...
\keyword{ network }