\name{print.summary.netlm}
\alias{print.summary.netlm}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ Printing for summary.netlm Objects }
\description{
  Prints an object of class \code{summary.netlm}.
}
\usage{
\method{print}{summary.netlm}(x, ...)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{x}{ An object of class \code{summary.netlm} }
  \item{\dots}{ Further arguments passed to or from other methods }
}
%\details{
%}
%\value{
%}
%\references{ }
\author{ Carter T. Butts \email{buttsc@uci.edu} }
%\note{  }

 %~Make other sections like WARNING with \section{WARNING }{....} ~

\seealso{ \code{\link{summary.netlm}} }

%\examples{
%
%}
\keyword{ print }%-- one or more ...
