\name{print.summary.netlogit}
\alias{print.summary.netlogit}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ Printing for summary.netlogit Objects }
\description{
  Prints an object of class \code{summary.netlogit}.
}
\usage{
\method{print}{summary.netlogit}(x, ...)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{x}{ An object of class \code{summary.netlogit}~ }
  \item{\dots}{ Further arguments passed to or from other methods }
}
%\details{
%}
%\value{
%}
%\references{  }
\author{ Carter T. Butts \email{buttsc@uci.edu} }
%\note{  }

 %~Make other sections like WARNING with \section{WARNING }{....} ~

\seealso{ \code{\link{summary.netlogit}} }

%\examples{
%
%}
\keyword{ print }%-- one or more ...
