\name{print.summary.qaptest}
\alias{print.summary.qaptest}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ Printing for summary.qaptest Objects  }
\description{
  Prints an object of class \code{summary.qaptest}.
}
\usage{
print.summary.qaptest(q)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{q}{ An object of class \code{summary.qaptest} }
}
%\details{
%}
\value{
}
%\references{  }
\author{ Carter T. Butts \email{ctb@andrew.cmu.edu} }
%\note{  }

 %~Make other sections like WARNING with \section{WARNING }{....} ~

\seealso{ \code{\link{summary.qaptest}} }

\examples{

}
\keyword{ print }%-- one or more ...
\keyword{ network }