\name{sna-deprecated}
\alias{sna-deprecated}
\alias{addisolates}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ Deprecated Functions in sna Package}
\description{
  These functions are provided for compatibility with older versions of \code{sna} only, and may be defunct as soon as the next release.
}
%\usage{
%addisolates(dat, n)
%}
%- maybe also `usage' for other objects documented here.
%\arguments{
%  \item{x}{ An object of class \code{cugtest} }
%  \item{\dots}{ Further arguments passed to or from other methods }
%}
\details{
The following \code{sna} functions are currently deprecated:

\begin{describe}
\item \code{addisolates}  
\end{describe}

The original help pages for these functions can be found at \code{help("oldName-deprecated")}.  Please avoid using them, since they will disappear....
}
%\value{
%None.
}
%\references{}
\author{ Carter T. Butts \email{buttsc@uci.edu} }
%\note{  }

% ~Make other sections like WARNING with \section{WARNING }{....} ~

\seealso{ \code{\link{Deprecated}} }

%\examples{
%}
\keyword{ misc }%-- one or more ...
