\name{summary.bayes.factor}
\alias{summary.bayes.factor}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ Detailed Summaries of Bayes Factor Objects }
\description{
Returns a \code{bayes.factor} summary object.
}
\usage{
summary.bayes.factor(bf)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{bf}{ An object of class \code{bayes.factor} }
}
%\details{
%}
\value{
An object of class \code{summary.bayes.factor}
}
%\references{  }
\author{ Carter T. Butts \email{ctb@andrew.cmu.edu} }
%\note{  }

 %~Make other sections like WARNING with \section{WARNING }{....} ~

\seealso{ \code{\link{bbnam.bf}} }

\examples{
}
\keyword{ network }%-- one or more ...
