\name{summary.bbnam}
\alias{summary.bbnam}
\alias{summary.bbnam.fixed}
\alias{summary.bbnam.pooled}
\alias{summary.bbnam.actor}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ Detailed Summaries of bbnam Objects }
\description{
  Returns a \code{bbnam} summary object
}
\usage{
summary.bbnam(b)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{b}{ An object of class \code{bbnam} }
}
%\details{
%}
\value{
An object of class \code{summary.bbnam}
}
%\references{  }
\author{ Carter T. Butts \email{ctb@andrew.cmu.edu} }
%\note{ }

 %~Make other sections like WARNING with \section{WARNING }{....} ~

\seealso{ \code{\link{bbnam}} }

\examples{

}
\keyword{ network }%-- one or more ...
