\name{summary.blockmodel}
\alias{summary.blockmodel}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ Detailed Summaries of blockmodel Objects }
\description{
  Returns a \code{blockmodel} summary object.
}
\usage{
summary.blockmodel(b)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{b}{ An object of class \code{blockmodel} }
}
%\details{
%
%}
\value{
An object of class \code{summary.blockmodel}
}
%\references{  }
\author{ Carter T. Butts \email{ctb@andrew.cmu.edu} }
%\note{ ~ }

 %~Make other sections like WARNING with \section{WARNING }{....} ~

\seealso{ \code{\link{blockmodel}} }

\examples{

}
\keyword{ network }%-- one or more ...
