\name{summary.cugtest}
\alias{summary.cugtest}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ Detailed Summaries of cugtest Objects }
\description{
  Returns a \code{cugtest} summary object
}
\usage{
summary.cugtest(c)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{c}{ An object of class \code{cugtest} }
}
%\details{
%
%}
\value{
An object of class \code{summary.cugtest}
}
%\references{ }
\author{ Carter T. Butts \email{ctb@andrew.cmu.edu} }
%\note{  }

 %~Make other sections like WARNING with \section{WARNING }{....} ~

\seealso{ \code{\link{cugtest}} }

\examples{

}
\keyword{ network }%-- one or more ...
