\name{summary.netcancor}
\alias{summary.netcancor}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ Detailed Summaries of netcancor Objects }
\description{
  Returns a \code{netcancor} summary object
}
\usage{
summary.netcancor(n)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{n}{ An object of class \code{netcancor} }
}
%\details{
%  
%}
\value{
  An object of class \code{summary.netcancor}
}
%\references{  }
\author{ Carter T. Butts \email{ctb@andrew.cmu.edu}~ }
%\note{  }

 %~Make other sections like WARNING with \section{WARNING }{....} ~

\seealso{ \code{\link{netcancor}} }

\examples{

}
\keyword{ network }%-- one or more ...
