\name{summary.netlm}
\alias{summary.netlm}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ Detailed Summaries of netlm Objects }
\description{
  Returns a \code{netlm} summary object
}
\usage{
summary.netlm(n)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{n}{ An object of class \code{netlm} }
}
%\details{
%  
%}
\value{
  An object of class \code{summary.netlm}
}
%\references{ }
\author{ Carter T. Butts \email{ctb@andrew.cmu.edu} }
%\note{  }

 %~Make other sections like WARNING with \section{WARNING }{....} ~

\seealso{ \code{\link{netlm}} }

\examples{

}
\keyword{ network }%-- one or more ...
