\name{summary.netlogit}
\alias{summary.netlogit}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ Detailed Summaries of netlogit Objects  }
\description{
  Returns a \code{netlogit} summary object~
}
\usage{
summary.netlogit(n)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{n}{  An object of class \code{netlogit}  }
}
%\details{
%}
\value{
  An object of class \code{summary.netlogit}
}
%\references{  }
\author{ Carter T. Butts \email{ctb@andrew.cmu.edu} }
%\note{  }

 %~Make other sections like WARNING with \section{WARNING }{....} ~

\seealso{\code{\link{netlogit}} }

\examples{

}
\keyword{ network }%-- one or more ...
