\name{summary.qaptest}
\alias{summary.qaptest}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ Detailed Summaries of qaptest Objects }
\description{
  Returns a \code{qaptest} summary object
}
\usage{
summary.qaptest(q)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{q}{ An object of class \code{qaptest} }
}
%\details{
%  
%}
\value{
  An object of class \code{summary.qaptest}
}
%\references{  }
\author{ Carter T. Butts \email{ctb@andrew.cmu.edu} }
%\note{  }

 %~Make other sections like WARNING with \section{WARNING }{....} ~

\seealso{ \code{\link{qaptest}} }

\examples{

}
\keyword{ network }%-- one or more ...
