\name{print.bayes.factor}
\alias{print.bayes.factor}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ Printing for Bayes Factor Objects }
\description{
  Prints a quick summary of a Bayes Factor object.
}
\usage{
\method{print}{bayes.factor}(x, ...)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{x}{ An object of class \code{bayes.factor} }
  \item{\dots}{ Further arguments passed to or from other methods }
}
%\details{
%}
\value{
None
}
%\references{  }
\author{ Carter T. Butts \email{buttsc@uci.edu} }
%\note{ }

% ~Make other sections like WARNING with \section{WARNING }{....} ~

\seealso{ \code{\link{bbnam.bf}}}

%\examples{
%}
\keyword{ print }%-- one or more ...
