\name{print.bbnam}
\alias{print.bbnam}
\alias{print.bbnam.fixed}
\alias{print.bbnam.pooled}
\alias{print.bbnam.actor}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ Printing for bbnam Objects }
\description{
Prints a quick summary of posterior draws from \code{\link{bbnam}}.
}
\usage{
\method{print}{bbnam}(x, ...)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{x}{ A \code{\link{bbnam}} object }
  \item{\dots}{ Further arguments passed to or from other methods }
}
%\details{
%}
\value{
None
}
%\references{ }
\author{ Carter T. Butts \email{buttsc@uci.edu} }
%\note{ }

% ~Make other sections like WARNING with \section{WARNING }{....} ~

\seealso{ \code{\link{bbnam}} }

%\examples{
%}
\keyword{ print }%-- one or more ...
