\name{print.netcancor}
\alias{print.netcancor}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ Printing for netcancor Objects }
\description{
  Prints a quick summary of objects produced by \code{\link{netcancor}}.
}
\usage{
\method{print}{netcancor}(x, ...)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{x}{ An object of class \code{netcancor}}
  \item{\dots}{ Further arguments passed to or from other methods }
}
%\details{
%}
%\value{
%}
%\references{}
\author{ Carter T. Butts \email{buttsc@uci.edu} }
%\note{  }

% ~Make other sections like WARNING with \section{WARNING }{....} ~

\seealso{ \code{\link{netcancor}} }

%\examples{
%}
\keyword{ print }%-- one or more ...
