\name{print.summary.bayes.factor}
\alias{print.summary.bayes.factor}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ Printing for summary.bayes.factor Objects }
\description{
  Prints an object of class \code{summary.bayes.factor}.
}
\usage{
\method{print}{summary.bayes.factor}(x, ...)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{x}{ An object of class \code{summary.bayes.factor} }
  \item{\dots}{ Further arguments passed to or from other methods }
}
%\details{
%}
%\value{
%}
%\references{  }
\author{ Carter T. Butts \email{buttsc@uci.edu} }
%\note{ }

% ~Make other sections like WARNING with \section{WARNING }{....} ~

\seealso{ \code{\link{summary.bayes.factor}} }

%\examples{
%}
\keyword{ print }%-- one or more ...
