\name{print.summary.bbnam}
\alias{print.summary.bbnam}
\alias{print.summary.bbnam.fixed}
\alias{print.summary.bbnam.pooled}
\alias{print.summary.bbnam.actor}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ Printing for summary.bbnam Objects }
\description{
  Prints an object of class \code{summary.bbnam}.
}
\usage{
\method{print}{summary.bbnam}(x, ...)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{x}{ An object of class \code{summary.bbnam} }
  \item{\dots}{ Further arguments passed to or from other methods }
}
%\details{
%}
%\value{
%}
%\references{ }
\author{ Carter T. Butts \email{buttsc@uci.edu} }
%\note{  }

% ~Make other sections like WARNING with \section{WARNING }{....} ~

\seealso{ \code{\link{bbnam}} }

%\examples{
%
%}
\keyword{ print }%-- one or more ...
