\name{print.summary.blockmodel}
\alias{print.summary.blockmodel}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ Printing for summary.blockmodel Objects }
\description{
  Prints an object of class \code{summary.blockmodel}.
}
\usage{
\method{print}{summary.blockmodel}(x, ...)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{x}{ An object of class \code{summary.blockmodel} }
  \item{\dots}{ Further arguments passed to or from other methods }
}
%\details{
%}
%\value{
%}
%\references{  }
\author{ Carter T. Butts \email{buttsc@uci.edu} }
%\note{ }

% ~Make other sections like WARNING with \section{WARNING }{....} ~

\seealso{ \code{\link{summary.blockmodel}}}

%\examples{
%
%}
\keyword{ print }%-- one or more ...
