\name{print.summary.cugtest}
\alias{print.summary.cugtest}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ Printing for summary.cugtest Objects }
\description{
  Prints an object of class \code{summary.cugtest}.
}
\usage{
\method{print}{summary.cugtest}(x, ...)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{x}{ An object of class \code{summary.cugtest} }
  \item{\dots}{ Further arguments passed to or from other methods }
}
%\details{
%}
%\value{
%}
%\references{  }
\author{ Carter T. Butts \email{buttsc@uci.edu} }
%\note{  }

% ~Make other sections like WARNING with \section{WARNING }{....} ~

\seealso{ \code{\link{summary.cugtest}} }

%\examples{
%
%}
\keyword{ print }%-- one or more ...
