\name{print.summary.qaptest}
\alias{print.summary.qaptest}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ Printing for summary.qaptest Objects  }
\description{
  Prints an object of class \code{summary.qaptest}.
}
\usage{
\method{print}{summary.qaptest}(x, ...)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{x}{ An object of class \code{summary.qaptest} }
  \item{\dots}{ Further arguments passed to or from other methods }
}
%\details{
%}
%\value{
%}
%\references{  }
\author{ Carter T. Butts \email{buttsc@uci.edu} }
%\note{  }

 %~Make other sections like WARNING with \section{WARNING }{....} ~

\seealso{ \code{\link{summary.qaptest}} }

%\examples{
%
%}
\keyword{ print }%-- one or more ...
