\name{sna-internal}
\alias{sna-internal}
\alias{bbnam.jntlik}
\alias{bbnam.jntlik.slice} 
\alias{bbnam.probtie}
\alias{logSum}
\alias{logMean}
\alias{logSub}
\alias{aggarray3d_R}
\alias{dyadcode_R}
\alias{logadd_R}
\alias{logsub_R}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ Internal sna Functions }
\description{
  Internal \code{sna} functions.
}
\usage{
bbnam.jntlik(dat, log=FALSE, ...)
bbnam.jntlik.slice(s, dat, a, em, ep, log=FALSE)
bbnam.probtie(dat, i, j, npriorij, em, ep)
logMean(x)
logSub(x, y)
logSum(x)
}
%- maybe also `usage' for other objects documented here.
\details{
These are not to be called by the end user.
}
\keyword{internal}
