\name{summary.bayes.factor}
\alias{summary.bayes.factor}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ Detailed Summaries of Bayes Factor Objects }
\description{
Returns a \code{bayes.factor} summary object.
}
\usage{
\method{summary}{bayes.factor}(object, ...)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{object}{ An object of class \code{bayes.factor} }
  \item{\dots}{ Further arguments passed to or from other methods }
}
%\details{
%}
\value{
An object of class \code{summary.bayes.factor}
}
%\references{  }
\author{ Carter T. Butts \email{buttsc@uci.edu} }
%\note{  }

 %~Make other sections like WARNING with \section{WARNING }{....} ~

\seealso{ \code{\link{bbnam.bf}} }

%\examples{
%}
\keyword{ math }%-- one or more ...
