\name{summary.bbnam}
\alias{summary.bbnam}
\alias{summary.bbnam.fixed}
\alias{summary.bbnam.pooled}
\alias{summary.bbnam.actor}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ Detailed Summaries of bbnam Objects }
\description{
  Returns a \code{bbnam} summary object
}
\usage{
\method{summary}{bbnam}(object, ...)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{object}{ An object of class \code{bbnam} }
  \item{\dots}{ Further arguments passed to or from other methods }
}
%\details{
%}
\value{
An object of class \code{summary.bbnam}
}
%\references{  }
\author{ Carter T. Butts \email{buttsc@uci.edu} }
%\note{ }

 %~Make other sections like WARNING with \section{WARNING }{....} ~

\seealso{ \code{\link{bbnam}} }

%\examples{
%
%}
\keyword{ math }%-- one or more ...
