\name{summary.blockmodel}
\alias{summary.blockmodel}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ Detailed Summaries of blockmodel Objects }
\description{
  Returns a \code{blockmodel} summary object.
}
\usage{
\method{summary}{blockmodel}(object, ...)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{object}{ An object of class \code{blockmodel} }
  \item{\dots}{ Further arguments passed to or from other methods }
}
%\details{
%
%}
\value{
An object of class \code{summary.blockmodel}
}
%\references{  }
\author{ Carter T. Butts \email{buttsc@uci.edu} }
%\note{ ~ }

 %~Make other sections like WARNING with \section{WARNING }{....} ~

\seealso{ \code{\link{blockmodel}} }

%\examples{
%
%}
\keyword{ math }%-- one or more ...
