\name{summary.cugtest}
\alias{summary.cugtest}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ Detailed Summaries of cugtest Objects }
\description{
  Returns a \code{cugtest} summary object
}
\usage{
\method{summary}{cugtest}(object, ...)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{object}{ An object of class \code{cugtest} }
  \item{\dots}{ Further arguments passed to or from other methods }
}
%\details{
%
%}
\value{
An object of class \code{summary.cugtest}
}
%\references{ }
\author{ Carter T. Butts \email{buttsc@uci.edu} }
%\note{  }

 %~Make other sections like WARNING with \section{WARNING }{....} ~

\seealso{ \code{\link{cugtest}} }

%\examples{
%
%}
\keyword{ math }%-- one or more ...
