\name{summary.netcancor}
\alias{summary.netcancor}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ Detailed Summaries of netcancor Objects }
\description{
  Returns a \code{netcancor} summary object
}
\usage{
\method{summary}{netcancor}(object, ...)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{object}{ An object of class \code{netcancor} }
  \item{\dots}{ Further arguments passed to or from other methods }
}
%\details{
%  
%}
\value{
  An object of class \code{summary.netcancor}
}
%\references{  }
\author{ Carter T. Butts \email{buttsc@uci.edu}~ }
%\note{  }

 %~Make other sections like WARNING with \section{WARNING }{....} ~

\seealso{ \code{\link{netcancor}} }

%\examples{
%
%}
\keyword{ math }%-- one or more ...
