\name{summary.netlm}
\alias{summary.netlm}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ Detailed Summaries of netlm Objects }
\description{
  Returns a \code{netlm} summary object
}
\usage{
\method{summary}{netlm}(object, ...)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{object}{ An object of class \code{netlm} }
  \item{\dots}{ Further arguments passed to or from other methods }
}
%\details{
%  
%}
\value{
  An object of class \code{summary.netlm}
}
%\references{ }
\author{ Carter T. Butts \email{buttsc@uci.edu} }
%\note{  }

 %~Make other sections like WARNING with \section{WARNING }{....} ~

\seealso{ \code{\link{netlm}} }

%\examples{
%
%}
\keyword{ math }%-- one or more ...
