\name{summary.netlogit}
\alias{summary.netlogit}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ Detailed Summaries of netlogit Objects  }
\description{
  Returns a \code{netlogit} summary object~
}
\usage{
\method{summary}{netlogit}(object, ...)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{object}{  An object of class \code{netlogit}  }
  \item{\dots}{ Further arguments passed to or from other methods }
}
%\details{
%}
\value{
  An object of class \code{summary.netlogit}
}
%\references{  }
\author{ Carter T. Butts \email{buttsc@uci.edu} }
%\note{  }

 %~Make other sections like WARNING with \section{WARNING }{....} ~

\seealso{\code{\link{netlogit}} }

%\examples{
%
%}
\keyword{ math }%-- one or more ...
