\name{summary.qaptest}
\alias{summary.qaptest}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ Detailed Summaries of qaptest Objects }
\description{
  Returns a \code{qaptest} summary object
}
\usage{
\method{summary}{qaptest}(object, ...)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{object}{ An object of class \code{qaptest} }
  \item{\dots}{ Further arguments passed to or from other methods }
}
%\details{
%  
%}
\value{
  An object of class \code{summary.qaptest}
}
%\references{  }
\author{ Carter T. Butts \email{buttsc@uci.edu} }
%\note{  }

 %~Make other sections like WARNING with \section{WARNING }{....} ~

\seealso{ \code{\link{qaptest}} }

%\examples{
%
%}
\keyword{ math }%-- one or more ...
